genai-security-ensemble/requirements.txt
sentence-transformers==2.7.0
transformers==4.45.2
scikit-learn==1.5.2
tabulate==0.9.0
fastapi==0.115.4
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
